Move widget API to separate file
